Fix configure with libSuperX11
Wine configure tests for library dependencies by compiling a test program and
checking that the library is a dependency of the binary. As the library linked
for libX* components of libSuperX11 is actually libX11.so.6, it fails to detect
and enable support for many X libraries (Xcomposite, Xi, Xrandr, Xrender etc).

This patch adds a fallback to the detection macro to check for libX11 when the
test program has successfully compiled and it hasn't detected a matching soname
dependency via readelf or ldd. Now all libX* components of libSuperX11 report
back as libX11.so.6 and enables their support in the wine build.

Requires latest autospec for it to set %reconfigure